+++ tests/config.py
@@ -104,6 +104,8 @@ class Config_(IntegrationSpec):
                     },
                     'tasks': {
                         'dedupe': True,
+                        'skip_pre': False,
+                        'skip_post': False,
                         'auto_dash_names': True,
                         'collection_name': 'tasks',
                         'search_root': None,